rename test classes to refer to the interface entrypoints
